backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
eventlet==0.33.3
Werkzeug==3.0.1

# PostgreSQL (solo para producción - comentado para desarrollo)
# psycopg2-binary==2.9.9

# Para producción con Gunicorn
# gunicorn==21.2.0